Hyperlink CHN  project page on the about page.
