+++ alderman_spending/lib/src/ui/faq_page/faq_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:alderman_spending/src/ui/navigation/navigation_drawer.dart';
-import 'package:url_launcher/url_launcher.dart';
+import 'package:alderman_spending/src/ui/faq_page/hyperlink.dart';
 
 class FAQScreen extends StatelessWidget {
   const FAQScreen({super.key});
@@ -380,42 +380,5 @@ class FAQTile extends StatelessWidget {
         subtitle: faq.answer,
       ),
     );
-  }
-}
-
-class HyperLink extends StatelessWidget {
-  final String displayText;
-  final String link;
-
-  HyperLink({required this.displayText, required this.link});
-
-  @override
-  Widget build(BuildContext context) {
-    return MouseRegion(
-      cursor: SystemMouseCursors.click,
-      child: GestureDetector(
-        onTap: () {
-          _launchURL(link);
-        },
-        child: Text(
-          displayText,
-          style: const TextStyle(
-            color: Colors.blue,
-            decoration: TextDecoration.underline,
-            decorationColor: Colors.blue,
-          ),
-        ),
-      ),
-    );
-  }
-
-  Future<void> _launchURL(String url) async {
-    final Uri _url = Uri.parse(url);
-
-    if (await canLaunchUrl(_url)) {
-      await launchUrl(_url);
-    } else {
-      throw 'Could not launch $url';
-    }
   }
 }
